+++ nodes/hm.nuke-workspace.py
@@ -40,6 +40,9 @@ nodes['hm.nuke-workspace'] = {
         'docker': {
             'enabled': True,
         },
+        'trivy': {
+            'enabled': True,
+        },
         'netplan': {
             'enabled': True,
             'ethernets': {